Splat. - More iso conversions for rendering - Move away from Vector2 towards integer coords - Playback speed should be parsed as an int - Cull our dumbass PUBLIC/PRIVATE macro - Start making methods under Manager classes static where possible - Change how loop works so mouse input no longer absorbs frames

+++ src/engine/EntityManager.cpp
@@ -100,7 +100,7 @@ void ct::EntityManager::Draw( const Camera &camera )
 
 void ct::EntityManager::SerializeEntities( Serializer *write )
 {
-	write->WriteI32( entities.size() );
+	write->WriteI32( ( int32_t ) entities.size() );
 
 	for ( auto &entity : entities )
 	{
